+++ src/main/java/com/trim/domain/member/entity/SocialType.java
@@ -1,5 +1,13 @@
 package com.trim.domain.member.entity;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
 public enum SocialType {
-    GOOGLE, NAVER, KAKAO
+    GOOGLE("google"), NAVER("naver"), KAKAO("kakao");
+
+    private final String key;
+
 }
