+++ src/main/java/com/trim/global/auth/service/CustomOAuth2UserService.java
@@ -28,23 +28,16 @@ import java.util.Optional;
 @Transactional
 public class CustomOAuth2UserService extends DefaultOAuth2UserService {
 
-    private final String TEMP_NICKNAME_HEADER = "TRIM_";
-    private final int TEMP_NICKNAME_LENGTH = 10;
-
-    private final NicknameService nicknameService;
-    private final MemberRepository memberRepository;
+    private final OAuth2MemberService oAuth2MemberService;
 
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
         OAuth2User oAuth2User = super.loadUser(userRequest);
         String registrationId = userRequest.getClientRegistration().getRegistrationId();
-        String userNameAttributeName = userRequest.getClientRegistration().getProviderDetails()
-                .getUserInfoEndpoint().getUserNameAttributeName();
         SocialType socialType = OAuth2Utils.getSocialType(registrationId);
 
-        log.info("registrationId={}", registrationId);
-        log.info("userNameAttributeName={}", userNameAttributeName);
-        log.info("socialType={}", socialType);
+        String userNameAttributeName = userRequest.getClientRegistration().getProviderDetails()
+                .getUserInfoEndpoint().getUserNameAttributeName();
 
         // 소셜에서 전달받은 정보를 가진 OAuth2User 에서 Map 을 추출하여 OAuth2Attribute 를 생성
         Map<String, Object> attributes = oAuth2User.getAttributes();
@@ -53,35 +46,12 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
         OAuth2Attributes oauth2Attributes = OAuth2Attributes.of(socialType, userNameAttributeName, attributes);
         OAuth2UserInfo oauth2UserInfo = oauth2Attributes.getOAuth2UserInfo();
 
-        // 값 추출
-        String socialId = oauth2UserInfo.getSocialId();
-        String email = oauth2UserInfo.getEmail();
-
-        log.info("socialId={}", socialId);
-        log.info("email={}", email);
-
-        String username = registrationId + "_" + socialId;
-        Optional<Member> targetMember = memberRepository.findByUsername(username);
-        Member member = targetMember.orElseGet(() -> signupSocialMember(username, email, socialType));
+        Member member = oAuth2MemberService.generateMemberFromOAuthUserInfo(oauth2UserInfo, socialType);
 
         return new CustomOAuthUser(member,
                 Collections.singleton(new SimpleGrantedAuthority(member.getRole().getKey())),
                 attributes);
     }
 
-    private Member signupSocialMember(String username, String email, SocialType socialType) {
-        String nickname = TEMP_NICKNAME_HEADER +
-                nicknameService.generateUniqueNickname(TEMP_NICKNAME_LENGTH);
-        // todo 닉네임 변경권 1개 지급
 
-        Member member = Member.builder()
-                .username(username)
-                .nickname(nickname)
-                .nicknameChangeChance(1)
-                .email(email)
-                .socialType(socialType)
-                .role(Role.USER)
-                .build();
-        return memberRepository.save(member);
-    }
 }
