--- src/main/java/com/trim/global/auth/service/NicknameService.java
@@ -1,23 +1,0 @@
-package com.trim.global.auth.service;
-
-import com.trim.domain.member.repository.MemberRepository;
-import com.trim.global.auth.utils.RandomStringGenerator;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class NicknameService {
-
-    private final MemberRepository memberRepository;
-
-    public String generateUniqueNickname(int length) {
-        String nickname;
-
-        do {
-            nickname = RandomStringGenerator.generateRandomString(length);
-        } while (memberRepository.existsByNickname(nickname));
-
-        return nickname;
-    }
-}
